Finally got A* working again
It took like 5 hours but I think I've gotten it pretty much fully functional again and have everything refactored to where it should be.

It turned out that the bug wasn't with the computePath method, it was a problem with the information I was feeding into it and that I was modifying it after calculating a path.